Add folders to separate systems
