+++ app/src/test/java/org/example/IngredientsTest.java
@@ -5,6 +5,8 @@ import java.time.LocalDate;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import org.example.backend_SSH.ingredient_logic.Ingredient;
 import org.junit.jupiter.api.Test;
 
 class IngredientsTest {
