+++ app/src/test/java/org/example/LoadRecipeTest.java
@@ -2,6 +2,9 @@ package org.example;
 
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import org.example.recipes_system.LoadRecipes;
+import org.example.recipes_system.Recipe;
 import org.junit.jupiter.api.Test;
 
 public class LoadRecipeTest {
