+++ app/src/test/java/org/example/TenantIngredientsTest.java
@@ -4,6 +4,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.util.List;
 
+import org.example.backend_SSH.ingredient_logic.MemoryIngredients;
 import org.junit.jupiter.api.Test;
 
 public class TenantIngredientsTest {
